+++ spec/choosy/printing/help_printer_spec.rb
@@ -9,11 +9,11 @@ module Choosy::Printing
       @c = Choosy::Command.new :foo do
         summary "This is a fairly long summary that should wrap around the whole screen at least once, so that I can test whether it's properly formatted"
 
-        header 'DESCRIPTION'
+        heading 'DESCRIPTION'
         para 'This is a description of this command that should span'
         para 'Multiple lines and carry itself beyond the average line length when actually called out from the unit tests itself so that we can correctly guage the line wrapping.'
 
-        header 'OPTIONS'
+        heading 'OPTIONS'
         boolean :evaluate, "The evaluation of some boolean something or other that really should span at least 3 lines of continuous text for testing the output of the option command."
         integer :count, "The count of something that should also really span multiple lines, if possible."
         boolean_ :debug, "Debug output"
@@ -29,7 +29,7 @@ module Choosy::Printing
       @b = @c.builder
 
       @s = Choosy::SuperCommand.new :super do
-        printer :standard, :color => true, :header_styles => [:bold, :blue]
+        printer :standard, :color => true, :heading_styles => [:bold, :blue]
         command @c
         metaname 'CMDS'